ECS header cleanup.
