+++ include/NovelRT/Ecs/Ecs.h
@@ -1,0 +1,15 @@
+// Copyright © Matt Jones and Contributors. Licensed under the MIT Licence (MIT). See LICENCE.md in the repository root for more information.
+
+#ifndef NOVELRT_ECS_H
+#define NOVELRT_ECS_H
+
+#include "Catalogue.h"
+#include "ComponentBuffer.h"
+#include "ComponentCache.h"
+#include "ComponentView.h"
+#include "EcsUtils.h"
+#include "EntityCache.h"
+#include "SparseSet.h"
+#include "SystemScheduler.h"
+
+#endif //!NOVELRT_ECS_H
